+++ Mods ChoGGi/Bottomless Waste Rock/metadata.lua
@@ -4,7 +4,7 @@ return PlaceObj("ModDef", {
 			"id", "ChoGGi_Library",
 			"title", "ChoGGi's Library",
 			"version_major", 6,
-			"version_minor", 4,
+			"version_minor", 6,
 		}),
 	},
 	"title", "Bottomless Waste Rock",