+++ Mods ChoGGi/Construction Show Map Sectors/Code/Script.lua
@@ -24,7 +24,8 @@ OnMsg.LoadGame = ModOptions
 local size = 100 * guim
 local green = green
 local MulDivRound = MulDivRound
-local PlaceObject = PlaceObject
+local IsValid = IsValid
+local type = type
 
 local orig_CursorBuilding_GameInit = CursorBuilding.GameInit
 function CursorBuilding.GameInit(...)
@@ -32,12 +33,13 @@ function CursorBuilding.GameInit(...)
 		return orig_CursorBuilding_GameInit(...)
 	end
 
+	local SectorUnexplored = SectorUnexplored
 	local g_MapSectors = g_MapSectors
 	for sector in pairs(g_MapSectors) do
 		if type(sector) == "table" then
 			sector.ChoGGi_decal = sector.decal
 			if not sector.decal then
-				sector.decal = PlaceObject("SectorUnexplored")
+				sector.decal = SectorUnexplored:new()
 				sector.decal:SetColorModifier(green)
 				sector.decal:SetPos(sector:GetPos())
 				sector.decal:SetScale(MulDivRound(sector.area:sizex(), 100, size)+1)
@@ -53,17 +55,18 @@ end
 
 local orig_CursorBuilding_Done = CursorBuilding.Done
 function CursorBuilding.Done(...)
+	if not mod_Option1 then
+		return orig_CursorBuilding_Done(...)
+	end
 
-	if mod_Option1 then
-		local g_MapSectors = g_MapSectors
-		for sector in pairs(g_MapSectors) do
-			if type(sector) == "table" then
-				if not sector.ChoGGi_decal then
-					sector.decal:delete()
-				end
-				sector.decal = sector.ChoGGi_decal
-				sector:UpdateDecal()
+	local g_MapSectors = g_MapSectors
+	for sector in pairs(g_MapSectors) do
+		if type(sector) == "table" then
+			if not sector.ChoGGi_decal then
+				sector.decal:delete()
 			end
+			sector.decal = sector.ChoGGi_decal
+			sector:UpdateDecal()
 		end
 	end